Scaling, Target and Multi-trenches e2e test
Check traffic behavior after target scale in/out.
Check if behavior of traffic and network interface after a target request/close a stream and after a target connect/disconnect a trench.
Tests if traffic can be sent to 2 trenches at the same time and using the same VIP address.
Tests if a target connected to a trench can be connected to another trench.