Identifier collision fix with registration process
Adding a new status to the target gives the possibility to have an extra
step in the target registration process.

A target can, first, register itself as 'disabled' to the NSP Service,
then fetch the targets, and verify if there is a collision with its
identifier or not. If there is, the target will generate a new
identifier and update its registration to the NSP service. If there is
no collision, the target will update its status to 'enabled' to the NSP
service.

The load-balancer will configure only the targets which are enabled, the
other ones will be ignored or removed if they were existing in the LB.

New defined interface for the Trench, Conduit and Stream (Separation in
different packages)

Add golang/mock dependency
mock nspAPI, conduit and trench to test the stream
Add '-race' and '-cover' to the make test command