Configuration manager on NSP service
- The configmap watcher has been re-implemented (pkg/monitor).
- The configmap Unmarshal functions and models has been updated with all
  resources and moved from the operator repository to this one
(pkg/reader).
- A memory registry has been implemented (pkg/registry).
- The configuration manager API has been implemented (pkg/manager).

When the configmap watcher/monitor will receive any event about the
changes from the kubernetes API, each yaml files contained in the
configmap will be unmarshalled and then converted and linked together
(e.g.: conduits referring to the trench...) with the nsp API types. All
resources will be registered in the configuration registry. On any
changes in the registry, an event is sent via the channel with the type
of resource modified. The registry also offers the possibility to get
any resource by filtering them with their parents and/or name, so the
manager server can return only the desired resources to the clients.