Fix egress traffic failover issue due to keepalive
Due to the new keepalive mechanism, the FE was refreshing its entry in
the NSP service every x (30) seconds, the NSP was receiving the changes
and was forwarding it to every component which uses the watch functions.
The LB were then receiving the changes about the FE every 30 seconds and
were doing some unecessary actions.

The ideal solution would be to have cache in the NSP, this will be
implemented in the future.