Allow internal target IP change without close/reopen
When a trench/conduit/stream is deleted and recreated, the target that
was connected will again request the connections with the old parameters
(IPs, GWs, routes...). The proxy might get different subnets allocated
and will forget the target IPs in use. So this will cause the targets to
have wrong/colliding IP addresses.

The proxy is now detecting if a target is requesting the connection with
invalid IP addresses and will fix them together with the other
parameters (GWs, routes...).

In addition, the tapa can now handle the internal IP changes and will
register itself with the new IP assigned.