Fix target changes on LB
If a target with an identifier was previously registered with IPs and is
now changing its IPs (IP refresh, different target...), the LB can now
handle that case by removing it completly and adding it again.

the internal setTargets is now removing old targets before adding new
ones. It also checks if IPs for an identifier have changed or not.