TAPA: fix concurrency issue in the stream retry
if multiple events occur (AddStream, SetStreams), a stack of Open would
pile up waiting for the mutex to be unlocked. When the mutex would be
unlocked (on close call), the close could then be executed, and after,
piled open calls would execute again, which would cause the stream to be
re-opened.