TAPA: fix conduit SetVips
SetVips could be called even after the conduit has been disconnected. So
the NSM connection would have been requested a new time and the
interface would be remaining in the pod.

The VIP watcher has been moved from the conduit handler to the conduit
manager (conduitRetry), so the SetVIPs won't be blocking any other
request and can be cancelled easily when disconnecting the conduit.