Fix configmap reconciler
The operator is reconciling the configmap all the time. It does that
because it compares the desire state with the current state, but the
changes between the states might only be ordering, so the configmap is
applied again and again. And that triggers the NSP events which forwards
the events to all other components (since we don't have cache) which
cause high CPU usage.
The configmap content is now sorted, so the configmap is reconciled
correctly.