TAPA; sync interface name cache for recovered connections
In case of a connection recovered through NSM's connection monitor,
the so called interface name cache in interfaceNameClient could have
become out of sync.
That's because the cache was not contacted in such cases, thus upon
a faulty Close() or TAPA crash the cache failed to keep track of
recovered names. (Which means, such names could be re-used by some other
connection.)

Therefor, move the interface name of recovered connections from Mechanism
to MechanismPreferences, so that intefaceNameClient could handle such
connection requests properly.
If the inteface name is not yet taken by some other connection, then use it,
and also update the inteface name cache if needed. (But generate a new name
if already taken by another connection.)