Use Prometheus exporter
Rename pkg metric to metrics.
Use observable counter instead of a counter since observable instrument
is more adapted for this case according to the open telemetry
documentation.
otpl exporter replaced by the prometheus via a http server.
Open Telemetry has to be updated to fix bugs related to temporality of
the metrics.