TAPA; gRPC health server hosts probe services
TAPA registers health services Liveness, Startup and Readiness,
which can be used by kubernetes probes to check the respective
status.

The evaluation criteria for these services can be changed behind
the secenes, without the need of updating the kubernetes probe
configuration of the TAPA container.

In accordance with the recent changes the example-target checks
Ambassador server availability as part of its liveness status.