e2e; phase out vlan-nsm based cluster breakout
Replace vlan-nsm with network-attachment-definition in e2e
test environment "kind-operator" to provide cluster breakout
functionality for Frontends.

Create additional docker network to be used for external
connectivity between Frontends and external-host containers.

The network-attachment-definition utilizes bridge CNI to create
the external interface in the FEs. While allocation of IP addresses
is achieved via Whereabouts IPAM CNI.
As part of the environment a VLAN aware Linux bridge is setup on each
Kind worker to be used by the bridge CNI.

Note: In order for BIRD routing suite (in Frontends) to be able to
get hold of a Router ID, the IPAM configuration is expected to assign
a feasible IPv4 address even in case of IPv6 only deployments.