ipam; introduce garbage collection logic
Updated db model with UpdatedAt and CreatedAt fields. Former one
is used to determine if a record associated with an NSM connection
is considered leftover and thus is subject to garbage collector based
clean up.

A one time clean up is also scheduled on startup to clean up records
created based on the old model that haven't been updated in the
meantime.