Update NSM API/SDK to use new scope selector
Use network service based scope selector limiting the
scope of the cluster connection monitor functionality
in stateless-lb. Thus, only NSM connection events that
are associated with TAPA->Proxy conns are forwarded to
stateless-lb in order to update its linux neighbor cache.

Go version had to be updated to 1.23 because of NSM.