operator; enable sequential conduit updates
Users can group Conduits into update groups by annotating
Conduit Custom Resources with the update-sync-group key.
Within an update group, Conduit updates are executed
sequentially rather than in parallel.

The annotation value of a Conduit Custom Resource can be
added/changed/removed on the fly.

The default annotation key can be changed via the operator
environment variable CONDUIT_UPDATE_SYNC_GROUP_KEY when
deploying the operator.

Locking is achieved using in-memory locks, which are fast
and provide a sufficiently good solution. However, this
approach is not effective in case of operator crashes
during ongoing updates. Such crashes would be considered a
bug to be fixed and should not occur in general.