tapa; Stream status reflects conduit connectivity
Stream status does not remain OPEN in case of connectivity
issues affecting the underlying NSM connection connecting
the TAPA with a local proxy.

The state changes can be observed via the TAP API.

Currently, NSP registration is not affected by the changes.

Note: It's a POC exploiting the fact that NSM datapath monitoring
is not enabled for the TAPA. Also, loss of the local nsmgr will
not trigger the Stream status update.